requirements.txt
fastapi
mysql_connector_repackaged
pydantic
uvicorn
mysql-connector-python
python-dotenv